+++ lib/ReactViews/Story/StoryBuilder.tsx
@@ -426,7 +426,7 @@ class StoryBuilder extends Component<
             />
             <Spacing right={1} />
             <TextSpan medium color={this.props.theme.infoColor}>
-              {t("story.saveInstructions")}
+              <Trans i18nKey="story.saveInstructions" />
             </TextSpan>
           </Box>
         )}